+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2025-09-05
+
 ### Changed
 
 - Upgrade `github.com/santhosh-tekuri/jsonschema` to v6
@@ -63,7 +65,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/giantswarm/schemadocs/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/schemadocs/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/giantswarm/schemadocs/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/giantswarm/schemadocs/compare/v0.0.7...v0.1.0
 [0.0.7]: https://github.com/giantswarm/schemadocs/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/giantswarm/schemadocs/compare/v0.0.5...v0.0.6